requirements.txt
# bottle==0.12.17
# The way Bottle currently handles error doesn't allow logging
# middleware to log them correctly, so used a patched version.
# NOTE: This won't override a previously installed version of Bottle,
#       so remove first with:
# 		> pip3 uninstall bottle
git+git://github.com/braedon/bottle@improve-wsgi-error-handling-012#egg=bottle
click==7.0
gevent==1.4.0
DBUtils==1.3
jog==0.1.1
PyMySQL==0.9.3
tonyg-rfc3339==0.1
